Add test045: BAOAB-limit Langevin dynamics
The accuracy of the BAOAB-limit discretization[1] of the over-damped
Langevin equation is compared against Euler-Maruyama discretization for
an artificial multi-well potential system.

The notebook shows an empirical estimation of the canonical distribution
computed from multiple simulation trajectories using each discretization
scheme. It is obvious that BAOAB-limit is better than Euler-Maruyama.

[1]: https://arxiv.org/abs/1203.5428