Add test143: fork(2) and exec(2)
I have no idea why (I could not find documentation), but when piping
into child process' stdin, the write-end of the pipe MUST be closed
before exec()-ing a command.

If the write-end is not closed, the child process blocks forever. I did
this mistake inside a loop and the OS died of resource exhaustion!