test155: Add test01:
The documentation suggests to add a unique tag for each point to avoid
points from overwritten when timestamps are the same[1]. I tried this
solution, and influxdb emits a warning after 100,000 insertions:

    WARN: 100% of max-values-per-tag limit exceeded: (100000/100000), db=testdb measurement=test tag=uniq service=store

The limit can be removed[2]. But it feels like a non-sustainable hack.
From my reading of the documentation[3] I understand that tags are not
meant to be used in this way. The issue that introduced the limit[4]
suggests that high cardinality data is bad. So, maybe I ought to try
another solution...

[1]: https://docs.influxdata.com/influxdb/v1.3/troubleshooting/frequently-asked-questions/#how-does-influxdb-handle-duplicate-points
[2]: https://stackoverflow.com/questions/43770354/max-values-per-tag-limit-exceeded-influxdb
[3]: https://docs.influxdata.com/influxdb/v1.3/concepts/schema_and_data_layout/#encode-meta-data-in-tags
[4]: https://github.com/influxdata/influxdb/issues/7146