test192: Fix hang up on applying firewall
Ansible's ssh connection is untracked until the firewall is applied. So
the existing connection gets blocked and Ansible hangs up. On the other
hand, new connections will be tracked and accepted by the firewall.

My solution is to apply the firewall asynchronously in "fire and forget"
manner. With a sufficient delay, subsequent tasks' connections should be
tracked and accepted by the firewall.

Another possible solution would be [1], but IMO async solution is
cleaner.

[1]: https://www.shysecurity.com/post/20160427-Ansible%20iptables%20workaround