test195: Remove Martian packets iptables rules
Spoofed packets are dropped by rp_filter[1]. Setting rp_filter to 1
enables full_check section in fib_validate_source()[2], which reverses
the source and the destination of a packet and tries a routing lookup[3].
So, rp_filter handles all the Martian packet rules. No need to filter
them with iptables.

[1]: https://www.kernel.org/doc/Documentation/networking/ip-sysctl.txt
[2]: https://github.com/torvalds/linux/blob/v4.14/net/ipv4/fib_frontend.c#L400
[3]: https://github.com/torvalds/linux/blob/v4.14/net/ipv4/fib_frontend.c#L329