Add test201: No-root-login Ubuntu on Vultr
The playbook adds a password-less sudoer to the system and disallows
root login. The sudoer can be tested with this command:

ansible -i admin_inventory --private-key path/to/privkey -bm ping all

admin_inventory is generated by the playbook and it allows the added
sudoer to run other playbooks.

Intended usage:

Provision, reboot and take a snapshot. Then use the snapshot as the base
image of other projects. So that Ansible can be run with the non-root
sudoer.