Add test481: Bolt DB for nonce verification
Bolt's API feels easier to use than LevelDB. And the transaction feature
looks right for the nonce verification job. Though the performance
characteristic of Bolt, which favors sequential reads to random writes,
is not the best match for nonce verification. However, I believe it's
okay since the nonce verification would not get that intensive.