Add test509: Send Go struct with AES-GCM
A message is comprised of three parts:

    messageType byte
    nonce       [12]byte
    ciphertext  [n]byte

The messageType part is authenticated with GCM as additional data. So a
possible man the middle cannot tamper with the message type.

The ciphertext part is the AES-GCM encryption of a plaintext using the
attached nonce. The plaintext is a JSON representation of a Go struct
representing a message.

I used JSON because it worked and was really easy to marshal/unmarshal
structs. I also tried ASN.1 DER but DER was too restrictive on types,
for example uint64 could not be handled. MessagePack (MP) could also be
a good low-overhead choice but the overhead of JSON compared to MP was
not that significant than that of AES-GCM encryption.

To prevent replay attacks a message must contain a counter. In this code
a counter is embedded in the struct to be marshaled and encrypted. The
counter could be placed in the plaintext message header like messageType
with authentication but it just complecated the implementation. Also
hiding the number of messages sent would be beneficial for preventing
some side-channel attacks.