Add test597: FAIL: setuid linuxbrew
setuid is ignored on shell scripts, so applying ug+s to the brew command
does not work. So I made a simple trampoline binary for setuid but it
seems euid and egid are reset to the real ones somewhere in the brew
command and the attempt did not work.

Anyway, `sudo -iu linuxbrew brew` work...