test901: Add forwarding/firewall rule
For some reason the port 6443 is inaccessible from Linux host:

    $ curl -k https://192.168.9.1:6443
    curl: (7) Failed to connect to 192.168.9.1 port 6443: Connection refused

While ping works:

    $ ping 192.168.9.1
    PING 192.168.9.1 (192.168.9.1) 56(84) bytes of data.
    64 bytes from 192.168.9.1: icmp_seq=1 ttl=64 time=0.034 ms
    64 bytes from 192.168.9.1: icmp_seq=2 ttl=64 time=0.043 ms
    ...

It works on mac host. Disabling ufw nor iptables does not help. So it should not
be a firewall problem. Can this be a Vagrant or VirtualBox problem?

I could make k3s accessible from host by addding port forwarding. But then the
packet is routed via eth0, not the expected eth1, and I cannot experiment with
the firewall on eth1.