Add test971: FreeBSD cron mail to local file
FreeBSD cron uses /usr/sbin/sendmail to send a mail. The program is
actually a mailwrapper and invokes a sendmail implementation listed in
/etc/mail/mailer.conf.

Here, we use a patched version of femtomail [1] as a sendmail. And
additional wrapper is required because femtomail does not interpret
sendmail -t option (find out recipients from To: header) set by
FreeBSD's cron.

[1] https://git.lekensteyn.nl/femtomail/