add test975: FreeBSD NFS, sticky bit
I set sticky bit to the shared directory and now Debian client can
create files on NFS shared volume as an arbitrary user without creating
the user on FreeBSD server.

For some reason GID on Debian is not "transferred" to FreeBSD and
"wheel" group is used instead. Only UID is transferred.