Add test981: Demonstrate signed integer overflow on multiplication
Someone says multiplying large signed integers produce negative
result due to signed integer overflow. But it does not always.
It is based on bit representation of the multiplied integers and
the highest bit of the product of the integers modulo 2^n where
n is the size of the integer in bits.