Add test997: docopt counting option (-v -vv -vvv ...)
Ugly. I expected the following format to work but did not.

  usage: main [options]

  options:
    -v   verbose
    -vv  more verbose

The supported formats look unfamiliar and thus are bad:

  main [-v ...]
  main [-v | -vv]

Maybe I'd resort to independent flags to mark different levels of
verbosity. Like this:

  options:
    -v  verbose
    -!  raise verbosity to the next level

So the user does this:

  $ main -v
  $ main -v!

No, this is strange. Do not do this. I really want -vv...